feat: move the search context hook
